requirements.txt
Flask==3.0.0
pandas==2.1.1
openpyxl==3.1.2
Flask-APScheduler==1.13.1
Jinja2==3.1.2
gunicorn==21.2.0
python-dateutil==2.8.2
